Updated 2 functions (Hover for details)
Changed the compare function so it doesn't hurt my soul anymore.
Updated the homework function to display a different message based on if a student finished their homework or just worked on it.
Removed the ability to leave the homework topic blank and just say "student worked on his/her homework".